V2016.8.5>>Update the code

+++ Modbus Poll/ModbusDllForVC.h
@@ -15,7 +15,7 @@ INPUT bool open_com(int m_com);
 INPUT void close_com();
 INPUT bool is_connect();
 INPUT int CheckTstatOnline(unsigned char devLo=1,unsigned char devHi=254);
-INPUT bool Change_BaudRate(unsigned short new_baurate);
+INPUT bool Change_BaudRate(int new_baurate);
 INPUT bool SetComm_Timeouts(LPCOMMTIMEOUTS lpCommTimeouts);
 INPUT void SetComnicationHandle(int nType,HANDLE hCommunication);
 INPUT bool Open_Socket(CString strIPAdress);
